chore(tocco-ui): separate styling from functionality
