+++ packages/tocco-ui/src/Button/index.js
@@ -1,3 +1,7 @@
-import Button, {StyledButton} from './Button'
+import Button from './Button'
+import StyledButton from './StyledButton'
 
-export {Button as default, StyledButton}
+export {
+  Button as default,
+  StyledButton
+}
