+++ packages/tocco-ui/src/ButtonGroup/index.js
@@ -1,3 +1,7 @@
-import ButtonGroup, {StyledButtonGroup} from './ButtonGroup'
+import ButtonGroup from './ButtonGroup'
+import StyledButtonGroup from './StyledButtonGroup'
 
-export {ButtonGroup as default, StyledButtonGroup}
+export {
+  ButtonGroup as default,
+  StyledButtonGroup
+}
