+++ packages/tocco-ui/src/ButtonLink/StyledButtonLink.js
@@ -1,0 +1,12 @@
+import {StyledButton} from '../Button'
+
+const StyledButtonLink = StyledButton.withComponent('a').extend`
+  && {
+    :hover,
+    :focus {
+      text-decoration: none;
+    }
+  }
+`
+
+export default StyledButtonLink
