+++ packages/tocco-ui/src/Menu/StyledItemAccordion.js
@@ -1,0 +1,3 @@
+import StyledItem from './StyledItem'
+
+export default StyledItem
