+++ packages/tocco-ui/src/Menu/StyledMenu.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+const StyledMenu = styled.ul`
+  && {
+    margin: 0;
+    padding: 0;
+    list-style: none;
+    /* width: fit-content; best option for dense mode but not supported for IE */
+  }
+`
+
+export default StyledMenu
