+++ packages/tocco-ui/src/Upload/index.js
@@ -1,3 +1,9 @@
+import StyledUploadInput from './StyledUploadInput'
+import StyledView from './StyledView'
 import Upload from './Upload'
 
-export default Upload
+export {
+  StyledUploadInput,
+  StyledView,
+  Upload as default
+}
